refactor: Paragraph to shared
