+++ src/features/certification/certification.tsx
@@ -1,6 +1,7 @@
 import classNames from 'classnames/bind';
-import { Paragraph, Title } from '@/app/components';
+import { Title } from '@/app/components';
 import { CourseNames } from '@/data/communication.data';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import styles from './certification.module.scss';
