+++ src/features/communication/communication.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable @stylistic/jsx-one-expression-per-line */
-import { Paragraph, Subtitle, Title } from '@/app/components';
+import { Subtitle, Title } from '@/app/components';
 import {
   CourseNames,
   DISCORD_LINKS,
@@ -9,6 +9,7 @@ import {
 import { TextLinkIcon } from '@/icons';
 import { DiscordLogo } from '@/icons/discord-logo';
 import { LinkCustom } from '@/shared/ui/link-custom';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './communication.scss';
