+++ src/features/home/about/about.tsx
@@ -1,7 +1,8 @@
-import { Paragraph, Subtitle, Title } from '@/app/components';
+import { Subtitle, Title } from '@/app/components';
 
 import image from '@/assets/about.webp';
 import Image from '@/features/image';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './about.scss';
