+++ src/features/home/merch/merch.tsx
@@ -1,10 +1,11 @@
-import { Paragraph, SectionLabel, Subtitle, Title } from '@/app/components';
+import { SectionLabel, Subtitle, Title } from '@/app/components';
 
 import { LINKS } from '@/app/const';
 import image from '@/assets/merch.webp';
 import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './merch.scss';
