+++ src/features/not-found/not-found.tsx
@@ -1,9 +1,9 @@
-import { Paragraph } from '@/app/components';
 import { useTitle } from '@/app/hooks';
 import notFoundImg from '@/assets/404.webp';
 import Image from '@/features/image';
 import { ArrowIcon } from '@/icons';
 import { LinkCustom } from '@/shared/ui/link-custom';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './not-found.scss';
