+++ src/features/school/about/about.tsx
@@ -1,4 +1,5 @@
-import { Paragraph, Subtitle, Title } from '@/app/components';
+import { Subtitle, Title } from '@/app/components';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './about.scss';
