+++ src/features/school/alumni/alumni.tsx
@@ -1,7 +1,8 @@
 import { alumni } from './constants';
-import { Paragraph, Title } from '@/app/components';
+import { Title } from '@/app/components';
 import { useWindowSize } from '@/app/hooks';
 import Image from '@/features/image';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './alumni.scss';
