+++ src/features/study-path/study-path.tsx
@@ -1,8 +1,9 @@
 import { createContext } from 'react';
 import { Stages } from './components';
-import { Paragraph, Title } from '@/app/components';
+import { Title } from '@/app/components';
 import { useDataByName } from '@/app/hooks';
 import { type DataMap } from '@/app/services/data/courses-data.types';
+import { Paragraph } from '@/shared/ui/paragraph';
 
 import './study-path.scss';
