+++ src/shared/ui/paragraph/paragraph.scss
